Example biomarker data

A dataset inspired by data collected by the Early Detection Research Network (EDRN). Biomarkers developed at six "labs" are validated at at least one of four "validation sites" on 306 cysts. The data also include two binary outcome variables: whether or not the cyst was classified as mucinous, and whether or not the cyst was determined to have high malignant potential.

biomarkers

Format

biomarkers: a tibble with 306 rows and 24 columns, where the first column is the validation site, the next two columns are the possible outcomes, and the remaining columns are the biomarkers:

institution

the validation site

mucinous

a binary indicator of whether the cyst was classified as mucinous

high_malignancy

a binary indicator of whether the cyst was classified as having high malignant potential

lab1_actb

a biomarker

lab1_molecules_score

a biomarker

lab1_telomerase_score

a biomarker

lab2_fluorescence_score

a biomarker

lab3_muc3ac_score

a biomarker

lab3_muc5ac_score

a biomarker

lab4_areg_score

a biomarker

lab4_glucose_score

a biomarker

lab5_mucinous_call

a biomarker (binary)

lab5_neoplasia_v1_call

a biomarker (binary)

lab5_neoplasia_v2_call

a biomarker (binary)

lab6_ab_score

a biomarker

cea

a biomarker

lab1_molecules_neoplasia_call

binary indicator of whether lab1_molecules_score > 25

lab1_telomerase_neoplasia_call

binary indicator of whether lab1_telomerase_score > 730

lab2_fluorescence_mucinous_call

binary indicator of whether lab2_fluorescence_score > 1.23

lab4_areg_mucinous_call

binary indicator of whether lab4_areg_score > 112

lab4_glucose_mucinous_call

binary indicator of whether lab4_glucose_score < 50

lab4_combined_mucinous_call

binary indicator of whether lab4_areg_score > 112 and lab4_glucose_score < 50

lab6_ab_neoplasia_call

binary indicator of whether lab6_ab_score > 0.104

cea_call

binary indicator of whether cea > 192

Source

Inspired by data collected by the EDRN https://edrn.nci.nih.gov/.