Get an augmented set based on the next-most significant variables

Based on the adjusted p-values from a FWER-controlling procedure and a more general error rate for which control is desired (e.g., generalized FWER, proportion of false positives, or FDR), augment the set based on FWER control with the next-most significant variables.

get_augmented_set(
  p_values = NULL,
  num_rejected = 0,
  alpha = 0.05,
  quantity = "gFWER",
  q = 0.05,
  k = 1
)

Arguments

p_values

the adjusted p-values.

num_rejected

the number of rejected null hypotheses from the base FWER-controlling procedure.

alpha

the significance level.

quantity

the quantity to control (i.e., "gFWER", "PFP", or "FDR").

q

the proportion for FDR or PFP control.

k

the number of false positives for gFWER control.

Value

a list of the variables selected into the augmentation set.