Get an initial selected set based on intrinsic importance and a base method

Using the estimated intrinsic importance and a base method designed to control the family-wise error rate (e.g., Holm), obtain an initial selected set.

get_base_set(
  test_statistics,
  p_values,
  alpha = 0.05,
  method = "maxT",
  B = 10000,
  Sigma = diag(1, nrow = length(test_statistics)),
  q = NULL
)

Arguments

test_statistics

the test statistics (used with "maxT")

p_values

(used with "minP" or "Holm")

alpha

the alpha level

method

the method (one of "maxT", "minP", or "Holm")

B

the number of resamples (for minP or maxT)

Sigma

the estimated covariance matrix for the test statistics

q

the false discovery rate (for method = "BY")

Value

the initial selected set