Control parameters for intrinsic variable selection

Control parameters for SPVIM-based intrinsic variable selection.

intrinsic_control(
  quantity = "gFWER",
  base_method = "Holm",
  fdr_method = "Holm",
  q = 0.2,
  k = 5
)

Arguments

quantity
the desired quantity for error-rate control: possible values are "gFWER" (the generalized family-wise error rate), "PFP" (the proportion of false positives), and "FDR" (the false discovery rate).

base_method
the family-wise error rate controlling method to use for obtaining the initial set of selected variables. Possible values are "maxT" and "minP" (for step-down procedures based on the test statistics ranked from largest to smallest or the p-values ranked from smallest to largest, respectively) or "Holm" for a procedure based on Holm-adjusted p-values.

fdr_method
the method for controlling the FDR (if quantity = "FDR"); possible values are "BY" (for Benjamini-Yekutieli) or one of the base_methods.

q
the desired proportion of false positives (only used if quantity = "PFP" or "FDR"; a fraction between 0 and 1).

k
the desired number of family-wise errors (an integer, greater than or equal to zero.)

Value

a list with the control parameters.